+++ prospector/__pkginfo__.py
@@ -1,5 +1,6 @@
 
-VERSION = (0, 5, 6)
+VERSION = (0, 7)
+
 
 def get_version():
     return '.'.join([str(v) for v in VERSION])