+++ prospector/tools/dodgy/__init__.py
@@ -1,8 +1,8 @@
 from __future__ import absolute_import
-
-from dodgy.run import run_checks
+import mimetypes
 import os
 import re
+from dodgy.run import check_file
 from prospector.message import Location, Message
 from prospector.tools.base import ToolBase
 
@@ -15,17 +15,27 @@ def module_from_path(path):
 
 class DodgyTool(ToolBase):
 
-    def prepare(self, rootpath, ignore, args, adaptors):
-        self.rootpath = rootpath
-        self.ignore = ignore
+    def prepare(self, found_files, args, adaptors):
+        self._files = found_files
 
     def run(self):
-        warnings = run_checks(self.rootpath, self.ignore)
-        messages = []
 
+        warnings = []
+        for filepath in self._files.iter_file_paths():
+            mimetype = mimetypes.guess_type(filepath)
+            if mimetype[0] is None or not mimetype[0].startswith('text/'):
+                continue
+            for line, code, message in check_file(filepath):
+                warnings.append({
+                    'line': line, 'code': code, 'message': message,
+                    'path': filepath
+                })
+
+        messages = []
         for warning in warnings:
             path = warning['path']
-            loc = Location(path, module_from_path(path), '', warning['line'], 0, absolute_path=False)
+            prefix = os.path.commonprefix([self._files.rootpath, path])
+            loc = Location(path, module_from_path(path[len(prefix):]), '', warning['line'], 0, absolute_path=True)
             msg = Message('dodgy', warning['code'], loc, warning['message'])
             messages.append(msg)