+++ prospector/tools/frosted/__init__.py
@@ -1,8 +1,6 @@
 from __future__ import absolute_import
 
-import os.path
-
-from frosted.api import iter_source_code, check_path
+from frosted.api import check_path
 
 from prospector.message import Location, Message
 from prospector.tools.base import ToolBase
@@ -74,10 +72,8 @@ class FrostedTool(ToolBase):
         self._paths = []
         self._ignores = []
 
-    def prepare(self, rootpath, ignore, args, adaptors):
-        self._paths = [rootpath]
-        self._rootpath = rootpath
-        self._ignores = ignore
+    def prepare(self, found_files, args, adaptors):
+        self._files = found_files
 
         for adaptor in adaptors:
             adaptor.adapt_frosted(self)
@@ -85,11 +81,17 @@ class FrostedTool(ToolBase):
     def run(self):
         reporter = ProspectorReporter(ignore=self.ignore_codes)
 
-        for filepath in iter_source_code(self._paths):
-            relpath = os.path.relpath(filepath, self._rootpath)
-            if any([ip.search(relpath) for ip in self._ignores]):
-                continue
-
-            check_path(filepath, reporter)
+        for filepath in self._files.iter_module_paths():
+            # Frosted cannot handle non-utf-8 encoded files at the moment -
+            # see https://github.com/timothycrosley/frosted/issues/53
+            # Therefore (since pyflakes overlaps heavily and does not have the same
+            # problem) we will simply suppress that error. If you do get it working
+            # correctly, you only end up with a "CannotDecodeFile" error anyway which
+            # is not useful to the user of prospector, nor is it actually a problem
+            # of the file but rather of frosted.
+            try:
+                check_path(filepath, reporter)
+            except UnicodeDecodeError:
+                pass
 
         return reporter.get_messages()