+++ prospector/tools/pep257/__init__.py
@@ -13,27 +13,14 @@ __all__ = (
 )
 
 
-def _find_code_files(rootpath, ignores):
-    code_files = []
-
-    for root, _, files in os.walk(rootpath):
-        for potential in files:
-            fullpath = os.path.join(root, potential)
-            relpath = os.path.relpath(fullpath, rootpath)
-            if potential.endswith('.py') and not any([ip.search(relpath) for ip in ignores]):
-                code_files.append(fullpath)
-
-    return code_files
-
-
 class Pep257Tool(ToolBase):
     def __init__(self, *args, **kwargs):
         super(Pep257Tool, self).__init__(*args, **kwargs)
         self._code_files = []
         self.ignore_codes = ()
 
-    def prepare(self, rootpath, ignore, args, adaptors):
-        self._code_files = _find_code_files(rootpath, ignore)
+    def prepare(self, found_files, args, adaptors):
+        self._code_files = list(found_files.iter_module_paths())
 
         for adaptor in adaptors:
             adaptor.adapt_pep257(self)