+++ prospector/tools/pyflakes/__init__.py
@@ -1,7 +1,6 @@
 from __future__ import absolute_import
-import os
 
-from pyflakes.api import iterSourceCode, checkPath
+from pyflakes.api import checkPath
 from pyflakes.reporter import Reporter
 
 from prospector.message import Location, Message
@@ -103,22 +102,15 @@ class PyFlakesTool(ToolBase):
         self._paths = []
         self._ignores = []
 
-    def prepare(self, rootpath, ignore, args, adaptors):
-        self._paths = [rootpath]
-        self._rootpath = rootpath
-        self._ignores = ignore
+    def prepare(self, found_files, args, adaptors):
+        self._files = found_files
 
         for adaptor in adaptors:
             adaptor.adapt_pyflakes(self)
 
     def run(self):
         reporter = ProspectorReporter(ignore=self.ignore_codes)
-
-        for filepath in iterSourceCode(self._paths):
-            relpath = os.path.relpath(filepath, self._rootpath)
-            if any([ip.search(relpath) for ip in self._ignores]):
-                continue
-
+        for filepath in self._files.iter_module_paths():
             checkPath(filepath, reporter)
 
         return reporter.get_messages()